Disables DocLint when compiling on Java 1.8 and higher.
DocLint is a new feature in Java 1.8, which provides a more strict checking of JavaDoc syntax.  Since we have a bunch of JavaDoc comments that aren't 100% in compliance with the strict syntax, we get DocLint build errors when we build on Java 1.8+.  By adding the flag: "-Xdoclint:none", we disable DocLint for Java 1.8.  However, since Java 1.7 and lower don't recognize this flag, we had to put the flag in a maven profile so it's not used when Java 1.7 and lower are detected.