Update AbstractChaiEntry to return null for 0
OpenDJ for some reason will save in the PWM passwordLastUpdateAttribute a 0
instead of leaving it null or empty string. When PWM tries to read a record
which has this value set, the help desk user or administrator has an error
displayed and the user logged out.

This also includes extra tests surrounding the date format string
including:
  * update comments for what a valid string is
  * include extra exception handling tests for EdirEntries
  * test for value is 0 and return null
  * update .gitignore for Intellij files to not be included
  * allow maven run the default maven test package since no tests
  * were being run anyway

Make EdirEntries convertZuluToDate testable

Since we are Creating a Calendar instance on each call,
we have 1000 possible dates as we are never setting the
millsecond field and as the zulu date format we are using
does not have milliseconds, then it should be 0 milliseconds
which makes testing if two dates are the same much simplier.