Support reproducible builds
META-INF/MANIFEST.MF contains values, that make the build harder
to reproduce than necessary.

Uses gitDescribe with tags, so that SCM-Git-Commit-ID-Description
contains the useful information, from which tag this build has
been built.

Removes SCM-Git-Branch, since this is the current HEAD branch.
When initially built, this is master, but when rebuilding,
master might have already moved along. The branch info
is actually not needed as long as SCM-Git-Commit-ID* properties
are provided.

Removes Implementation-Build-Java-Vendor and
Implementation-Build-Java-Version as these are too specific
(e.g. AdoptOpenJDK vs. OpenJDK vs. Oracle and 14.0.1 vs. 14.0.2).
Instead the property Implementation-Build-Java-Spec-Version is
used.